django-issuer-platform/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
dj-database-url==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
web3==6.13.0
eth-account==0.10.0
celery==5.3.4
django-celery-beat==2.5.0
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
cryptography==41.0.7
requests==2.31.0
ipfshttpclient==0.8.0a2
